Pull shared code into a module
This implements @glennw's PR #5 to move the hashing code used by both
phf and phf_mac into a module imported by both crates. It breaks the
hard dependency of phf_mac on phf, which means phf doesn't have to be
compiled for the host architecture as well as the target. It also means
we can have a dev-dependency from phf to phf_mac for tests.