Add hash() and get_index() to phf_shared.
In https://github.com/servo/string-cache, we currently use a `phf::OrderedSet`
with its `get_index` method to get an identified stored in an `Atom`,
and `index` to get a string back from that identifier.

However, the extra inderection of `OrderedSet` of `Set` is not necessary.
We don’t care about the order, only about getting numeric identifiers.

Additionally, when `get_index` returns `None`,
we hash the input string again to find it in table of dynamic atoms.
With this chang, we can reuse the phf hash instead:

https://github.com/servo/string-cache/pull/103

At first I tried adding hash and index access to `phf::Map`,
but the API got messy quickly.