Remove PhfHash in favor of AsRef<[u8]>
This opens the door to different implementations, like one that
dynamically generates a hash function and doesn't have to look at all of
the bytes. This removes support for most of the types, but we really
want to focus on strings and slices. Normal matches work just fine on
integral types.